--- a/src/octoprint/settings.py
+++ b/src/octoprint/settings.py
@@ -645,6 +645,8 @@ class Settings(object):
 		self._config = None
 		self._dirty = False
 		self._dirty_time = 0
+		self._last_config_hash = None
+		self._last_effective_hash = None
 		self._mtime = None
 
 		self._get_preprocessors = dict(
@@ -842,6 +844,15 @@ class Settings(object):
 
 		return list(map(process_control, controls))
 
+	def _forget_hashes(self):
+		self._last_config_hash = None
+		self._last_effective_hash = None
+
+	def _mark_dirty(self):
+		self._dirty = True
+		self._dirty_time = time.time()
+		self._forget_hashes()
+
 	@property
 	def effective(self):
 		return self._map.deep_dict()
@@ -853,10 +864,14 @@ class Settings(object):
 
 	@property
 	def effective_hash(self):
+		if self._last_effective_hash is not None:
+			return self._last_effective_hash
+
 		import hashlib
 		hash = hashlib.md5()
 		hash.update(self.effective_yaml.encode('utf-8'))
-		return hash.hexdigest()
+		self._last_effective_hash = hash.hexdigest()
+		return self._last_effective_hash
 
 	@property
 	def config_yaml(self):
@@ -865,10 +880,14 @@ class Settings(object):
 
 	@property
 	def config_hash(self):
+		if self._last_config_hash:
+			return self._last_config_hash
+
 		import hashlib
 		hash = hashlib.md5()
 		hash.update(self.config_yaml.encode('utf-8'))
-		return hash.hexdigest()
+		self._last_config_hash = hash.hexdigest()
+		return self._last_config_hash
 
 	@property
 	def _config(self):
@@ -932,6 +951,8 @@ class Settings(object):
 
 		if migrate:
 			self._migrate_config()
+
+		self._forget_hashes()
 
 	def load_overlay(self, overlay, migrate=True):
 		config = None
@@ -1603,8 +1624,7 @@ class Settings(object):
 					del self._config["folder"][type]
 					if not len(self._config["folder"]):
 						del self._config["folder"]
-					self._dirty = True
-					self._dirty_time = time.time()
+					self._mark_dirty()
 					self.save()
 				except KeyError:
 					pass
@@ -1652,8 +1672,7 @@ class Settings(object):
 
 		try:
 			chain.del_by_path(path)
-			self._dirty = True
-			self._dirty_time = time.time()
+			self._mark_dirty()
 		except KeyError:
 			if error_on_path:
 				raise NoSuchSettingsPath()
@@ -1711,8 +1730,7 @@ class Settings(object):
 		if not force and in_defaults and in_local and default_value == value:
 			try:
 				chain.del_by_path(path)
-				self._dirty = True
-				self._dirty_time = time.time()
+				self._mark_dirty()
 			except KeyError:
 				if error_on_path:
 					raise NoSuchSettingsPath()
@@ -1722,8 +1740,7 @@ class Settings(object):
 				chain.del_by_path(path)
 			else:
 				chain.set_by_path(path, value)
-			self._dirty = True
-			self._dirty_time = time.time()
+			self._mark_dirty()
 
 	def setInt(self, path, value, **kwargs):
 		if value is None:
@@ -1785,8 +1802,7 @@ class Settings(object):
 			del self._config["folder"][type]
 			if not self._config["folder"]:
 				del self._config["folder"]
-			self._dirty = True
-			self._dirty_time = time.time()
+			self._mark_dirty()
 		elif (path != currentPath and path != defaultPath) or force:
 			if validate:
 				_validate_folder(path, check_writable=True, deep_check_writable=True)
@@ -1794,8 +1810,7 @@ class Settings(object):
 			if "folder" not in self._config:
 				self._config["folder"] = {}
 			self._config["folder"][type] = path
-			self._dirty = True
-			self._dirty_time = time.time()
+			self._mark_dirty()
 
 	def saveScript(self, script_type, name, script):
 		script_folder = self.getBaseFolder("scripts")