@foosel 2020-04-06
Manually merged into `maintenance` (realized too late that this was targeting `devel` and should have been rebased first)